feat: added update_phar
