+++ .idea/sqldialects.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="SqlDialectMappings">
     <file url="file://$PROJECT_DIR$/notatki2.py" dialect="PostgreSQL" />
-    <file url="file://$PROJECT_DIR$/utils/crud.py" dialect="GenericSQL" />
+    <file url="file://$PROJECT_DIR$/utils/crud_phar.py" dialect="GenericSQL" />
     <file url="PROJECT" dialect="PostgreSQL" />
   </component>
 </project>
