+++ main.py
@@ -1,4 +1,4 @@
-from notatki2 import create_phar, read_phar,db_params
+from utils.crud_phar import create_phar, read_phar,db_params,remove_phar,update_phar
 if __name__ == '__main__':
     print('Witaj użytkowniku! ')
     while True:
@@ -7,7 +7,7 @@ if __name__ == '__main__':
         print('1. Pokaż listę aptek: ')
         print('2. Pokaż listę leków: ')
         print('3. Pokaż listę pacjentów: ')
-        menu_option: str = input('Wybierz dostępną funckję z menu: ')
+        menu_option: str = input('Wybierz dostępną funkcję z menu: ')
         if menu_option == '0':
             break
         if menu_option == '1':
@@ -16,12 +16,15 @@ if __name__ == '__main__':
             print('1. Edytuj dane apteki: ')
             print('2. Dodaj nową aptekę do listy: ')
             print('3. Usuń aptekę z listy: ')
-            sub_menu1_option: str = input('Wybierz dostępną funckję z menu: ')
+            sub_menu1_option: str = input('Wybierz dostępną funkcję z menu: ')
             if sub_menu1_option == '0':
                 back_to_main_menu = True
 
-            #if sub_menu1_option == '1':
-
+            if sub_menu1_option == '1':
+                update_phar(db_params)
+                print('Edytowano. ')
             if sub_menu1_option == '2':
                 create_phar(db_params)
-            #if sub_menu1_option == '3':
+            if sub_menu1_option == '3':
+                remove_phar(db_params)
+                print('Usunięto.')
